memory leak fixes * Added a bunch of simple leak detection checks to the tests. * Fixed some real leaks in slotmap and thread * Fixed a bug with allocating memory of large alignments * Fixed some leaks in test code

+++ rjd_hash.h
@@ -60,8 +60,8 @@ struct rjd_hash32 rjd_hash32_data(const uint8_t* key, int length)
 		return hash;
 	}
 
-	const uint32_t PRIME = 16777619;
-	const uint32_t SEED  = 2166136261;
+	const uint64_t PRIME = 16777619;
+	const uint64_t SEED  = 2166136261;
 
 	struct rjd_hash32 hash = { SEED };
 	if (length == -1) {
@@ -72,7 +72,7 @@ struct rjd_hash32 rjd_hash32_data(const uint8_t* key, int length)
 		while (length > 0)
 		{
 			--length;
-			hash.value = (*key++ ^ hash.value) * PRIME;
+			hash.value = (uint32_t)((*key++ ^ hash.value) * PRIME);
 		}
 	}
 	return hash;
